twitter call now moved to profile controller
