+++ config/twitter-connection.js
@@ -9,26 +9,4 @@ var twitterClient = new Twitter({
     bearer_token: process.env.TW_BEARER_TOKEN,
 });
 
-var params = { screen_name: "nodejs" };
-
-async function twitterConnection(paramsObject) {
-    twitterClient.get(
-        "statuses/user_timeline",
-        paramsObject,
-        async function (error, tweets, response) {
-            if (!error) {
-                const test = await response;
-                myTest(test);
-            }
-        }
-    );
-}
-
-function myTest(test) {
-    var testr = JSON.parse(test.body);
-    console.log(testr[0]);
-}
-
-twitterConnection(params);
-
 module.exports = twitterClient;
