+++ server.js
@@ -7,7 +7,7 @@ const session = require("express-session");
 const path = require("path");
 const SequelizeStore = require("connect-session-sequelize")(session.Store);
 const sequelize = require("./config/connection");
-const twitterConnection = require("./config/twitter-connection");
+// const twitterConnection = require("./config/twitter-connection");
 const controllers = require("./controllers");
 const helpers = require("./utils/helpers");
 const { errorMonitor } = require("events");
